+++ wait-serv.py
@@ -1,0 +1,3 @@
+from waitress import serve
+import kkbridge
+serve(kkbridge.create_app(), host='0.0.0.0', port=5000)
